chore(readability): use worker threads
happy-dom throw unhandled exceptions based on the target URL markup, so there is no way to don't pullate the main Node.js process.
