+++ packages/metascraper-readability/package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "@metascraper/helpers": "workspace:*",
     "@mozilla/readability": "~0.5.0",
+    "async-memoize-one": "~1.1.8",
     "happy-dom": "~16.5.3"
   },
   "devDependencies": {
@@ -37,7 +38,10 @@
     "src"
   ],
   "scripts": {
-    "test": "NODE_PATH=.. TZ=UTC ava --timeout 15s"
+    "test": "NODE_PATH=.. TZ=UTC ava"
   },
-  "license": "MIT"
+  "license": "MIT",
+  "ava": {
+    "timeout": "15s"
+  }
 }
