+++ packages/metascraper-readability/src/index.d.ts
@@ -1,5 +1,4 @@
 type Options = {
-  getDocument: ({url: string, html: string }) => Document,
   readabilityOpts: import('readability').ReadabilityOptions,
 }
